Add setup test w/ age and git
Fixes #2108

RELEASE_NOTES=n/a
